Break out step_definitions into strategies for finding and creating attributes
The change 1b28476713df98bed3963521e200ed34485a4209 introduced an issue
where projects not using ActiveRecord would fail because the constant
wasn't defined.

This removes the reliance on ActiveRecord and uses different strategies
for building hashes from cucumber tables based on whether the attributes
are finding a record in the database or creating an instance with
factory girl.