Prefer string argument to class option [ci skip]
This swaps the string and constant examples for the `class:` option,
treating the string example as the normal case and the constant example
as an alternate approach with caveats.

I have seen a number of problems with using constants, and
[rubocop-rspec][] even has a custom cop to prevent this.

I could almost remove the constant example, except that a lot of folks
are still using it and there may be room for confusion if it totally
disappears. Instead, I present it with a note about the potential
performance problem.

[rubocop-rspec]: https://github.com/rubocop-hq/rubocop-rspec/blob/master/lib/rubocop/cop/rspec/factory_bot/factory_class_name.rb